(svn r20722) -Fix (r20719): signed/unsigned side effect makes a small negative value become a big positive value

+++ src/statusbar_gui.cpp
@@ -182,7 +182,7 @@ struct StatusBarWindow : Window {
 
 				if (this->reminder_timeout > 0) {
 					Dimension icon_size = GetSpriteSize(SPR_UNREAD_NEWS);
-					DrawSprite(SPR_UNREAD_NEWS, PAL_NONE, r.right - WD_FRAMERECT_RIGHT - icon_size.width, r.top + WD_FRAMERECT_TOP + (FONT_HEIGHT_NORMAL - icon_size.height) / 2);
+					DrawSprite(SPR_UNREAD_NEWS, PAL_NONE, r.right - WD_FRAMERECT_RIGHT - icon_size.width, r.top + WD_FRAMERECT_TOP + (int)(FONT_HEIGHT_NORMAL - icon_size.height) / 2);
 				}
 				break;
 		}
